+++ src/User/UserAttributeItemProviderInterface.php
@@ -7,14 +7,15 @@ namespace Dbp\Relay\CoreBundle\User;
 /**
  * Extended version of UserAttributeProviderInterface which also allows retrieving single attributes.
  */
-interface UserAttributeProviderExInterface extends UserAttributeProviderInterface
+interface UserAttributeItemProviderInterface extends UserAttributeProviderInterface
 {
     /**
-     * Returns the attribute value for the given $userIdentifier.
+     * Should return the user attribute value for the user with the given $userIdentifier.
      *
      * In case $userIdentifier is null, the default values of the attribute should be returned.
      *
-     * If the passed $name is not known it should throw UserAttributeException.
+     * If the user attribute with the given $name is not defined for this provider,
+     * i.e., hasUserAttribute($name) returns false, it should throw UserAttributeException.
      *
      * @return array<string, mixed>
      *
@@ -23,7 +24,8 @@ interface UserAttributeProviderExInterface extends UserAttributeProviderInterfac
     public function getUserAttribute(?string $userIdentifier, string $name): mixed;
 
     /**
-     * Returns whether the given attribute name is available.
+     * Should return true if the given user attribute $name is defined for this provider,
+     * i.e., getUserAttribute will return a value for this attribute, false otherwise.
      */
     public function hasUserAttribute(string $name): bool;
 }