+++ src/User/UserAttributeMuxer.php
@@ -13,14 +13,17 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
  */
 class UserAttributeMuxer
 {
-    /** @var iterable<UserAttributeProviderInterface> */
-    private iterable $userAttributeProviders;
+    /** @var UserAttributeCollectionProviderInterface[] */
+    private array $userAttributeCollectionProviders = [];
+
+    /** @var UserAttributeItemProviderInterface[] */
+    private array $userAttributeItemProviders = [];
 
     /** @var array<int, array> */
-    private array $providerCache = [];
+    private array $valuesPerCollectionProviderCache = [];
 
     /** @var array<int, string[]> */
-    private array $availableCache = [];
+    private array $availableAttributesPerCollectionProviderCache = [];
 
     /** @var string[] */
     private array $attributeStack = [];
@@ -28,13 +31,29 @@ class UserAttributeMuxer
     /**
      * @var ?string[]
      */
-    private ?array $availableCacheAll = null;
+    private ?array $additionallyAvailableAttributesCache = null;
 
     public function __construct(
         UserAttributeProviderProviderInterface $userAttributeProviderProvider,
         private readonly EventDispatcherInterface $eventDispatcher)
     {
-        $this->userAttributeProviders = $userAttributeProviderProvider->getAuthorizationDataProviders();
+        foreach ($userAttributeProviderProvider->getAuthorizationDataProviders() as $userAttributeProvider) {
+            if ($userAttributeProvider instanceof UserAttributeCollectionProviderInterface) {
+                $this->userAttributeCollectionProviders[] = $userAttributeProvider;
+            } elseif ($userAttributeProvider instanceof UserAttributeItemProviderInterface) {
+                $this->userAttributeItemProviders[] = $userAttributeProvider;
+            }
+        }
+    }
+
+    /**
+     * Clears all caches. For testing purposes.
+     */
+    public function clearCaches(): void
+    {
+        $this->additionallyAvailableAttributesCache = null;
+        $this->valuesPerCollectionProviderCache = [];
+        $this->availableAttributesPerCollectionProviderCache = [];
     }
 
     /**
@@ -44,70 +63,14 @@ class UserAttributeMuxer
      */
     public function getAvailableAttributes(): array
     {
-        if ($this->availableCacheAll === null) {
-            $res = [];
-            foreach ($this->userAttributeProviders as $userAttributeProvider) {
-                $availableAttributes = $this->getProviderAvailableAttributes($userAttributeProvider);
-                $res = array_merge($res, $availableAttributes);
-            }
+        $availableAttributes = $this->getAdditionallyAvailableAttributesCached();
 
-            $event = new GetAvailableUserAttributesEvent($res);
-            $this->eventDispatcher->dispatch($event);
-            $this->availableCacheAll = $event->getAttributes();
+        foreach ($this->userAttributeCollectionProviders as $userAttributeProvider) {
+            $availableAttributes = array_merge($availableAttributes,
+                $this->getProviderAvailableAttributesCached($userAttributeProvider));
         }
 
-        return $this->availableCacheAll;
-    }
-
-    /**
-     * Returns a cached list for available attributes for the provider.
-     *
-     * @return string[]
-     */
-    private function getProviderAvailableAttributes(UserAttributeProviderInterface $prov): array
-    {
-        // Caches getAvailableAttributes for each provider
-        $provKey = spl_object_id($prov);
-        if (!array_key_exists($provKey, $this->availableCache)) {
-            $this->availableCache[$provKey] = $prov->getAvailableAttributes();
-        }
-
-        return $this->availableCache[$provKey];
-    }
-
-    /**
-     * Returns a cached map of available user attributes.
-     *
-     * @return array<string, mixed>
-     */
-    private function getProviderUserAttributes(UserAttributeProviderInterface $userAttributeProvider, ?string $userIdentifier): array
-    {
-        // We cache the attributes for each provider, but only for the last userIdentifier
-        $provKey = spl_object_id($userAttributeProvider);
-        if (!array_key_exists($provKey, $this->providerCache) || $this->providerCache[$provKey][0] !== $userIdentifier) {
-            $this->providerCache[$provKey] = [$userIdentifier, $userAttributeProvider->getUserAttributes($userIdentifier)];
-        }
-        $res = $this->providerCache[$provKey];
-        assert($res[0] === $userIdentifier);
-
-        return $res[1];
-    }
-
-    /**
-     * @throws UserAttributeException
-     */
-    private function getProviderUserAttribute(UserAttributeProviderInterface $userAttributeProvider, ?string $userIdentifier, string $attributeName): mixed
-    {
-        if ($userAttributeProvider instanceof UserAttributeProviderExInterface) {
-            return $userAttributeProvider->getUserAttribute($userIdentifier, $attributeName);
-        } else {
-            $userAttributes = $this->getProviderUserAttributes($userAttributeProvider, $userIdentifier);
-            if (!array_key_exists($attributeName, $userAttributes)) {
-                throw new UserAttributeException(sprintf('attribute \'%s\' was not provided', $attributeName), UserAttributeException::USER_ATTRIBUTE_UNDEFINED);
-            }
-
-            return $userAttributes[$attributeName];
-        }
+        return $availableAttributes;
     }
 
     /**
@@ -115,18 +78,35 @@ class UserAttributeMuxer
      */
     public function getAttribute(?string $userIdentifier, string $attributeName, mixed $defaultValue = null): mixed
     {
-        if (!in_array($attributeName, $this->getAvailableAttributes(), true)) {
-            throw new UserAttributeException(sprintf('attribute \'%s\' undefined', $attributeName), UserAttributeException::USER_ATTRIBUTE_UNDEFINED);
+        $value = null;
+        $found = false;
+
+        if (in_array($attributeName, $this->getAdditionallyAvailableAttributesCached(), true)) {
+            $found = true;
+        } else {
+            foreach ($this->userAttributeCollectionProviders as $userAttributeCollectionProvider) {
+                if (in_array($attributeName,
+                    $this->getProviderAvailableAttributesCached($userAttributeCollectionProvider), true)) {
+                    $value = $this->getProviderUserAttribute($userAttributeCollectionProvider, $userIdentifier, $attributeName);
+                    $found = true;
+                    break;
+                }
+            }
+
+            if (false === $found) {
+                foreach ($this->userAttributeItemProviders as $userAttributeItemProvider) {
+                    if ($userAttributeItemProvider->hasUserAttribute($attributeName)) {
+                        $value = $userAttributeItemProvider->getUserAttribute($userIdentifier, $attributeName);
+                        $found = true;
+                        break;
+                    }
+                }
+            }
         }
 
-        $value = null;
-        foreach ($this->userAttributeProviders as $authorizationDataProvider) {
-            $availableAttributes = $this->getProviderAvailableAttributes($authorizationDataProvider);
-            if (!in_array($attributeName, $availableAttributes, true)) {
-                continue;
-            }
-            $value = $this->getProviderUserAttribute($authorizationDataProvider, $userIdentifier, $attributeName);
-            break;
+        if (false === $found) {
+            throw new UserAttributeException(sprintf('attribute \'%s\' undefined', $attributeName),
+                UserAttributeException::USER_ATTRIBUTE_UNDEFINED);
         }
 
         $event = new GetUserAttributeEvent($this, $attributeName, $value, $userIdentifier);
@@ -137,7 +117,7 @@ class UserAttributeMuxer
             throw new UserAttributeException(sprintf('infinite loop caused by a %s subscriber. authorization attribute: %s',
                 GetUserAttributeEvent::class, $attributeName), UserAttributeException::INFINITE_EVENT_LOOP_DETECTED);
         }
-        array_push($this->attributeStack, $attributeName);
+        $this->attributeStack[] = $attributeName;
         try {
             $this->eventDispatcher->dispatch($event);
         } finally {
@@ -145,5 +125,72 @@ class UserAttributeMuxer
         }
 
         return $event->getAttributeValue() ?? $defaultValue;
+    }
+
+    /**
+     * Returns a cached list for available attributes for the provider.
+     *
+     * @return string[]
+     */
+    private function getProviderAvailableAttributesCached(
+        UserAttributeCollectionProviderInterface $userAttributeCollectionProvider): array
+    {
+        // Caches getAvailableAttributes for each provider
+        $providerKey = spl_object_id($userAttributeCollectionProvider);
+        if (false === array_key_exists($providerKey, $this->availableAttributesPerCollectionProviderCache)) {
+            $this->availableAttributesPerCollectionProviderCache[$providerKey] = $userAttributeCollectionProvider->getAvailableAttributes();
+        }
+
+        return $this->availableAttributesPerCollectionProviderCache[$providerKey];
+    }
+
+    /**
+     * Returns a cached map of available user attributes.
+     *
+     * @return array<string, mixed>
+     */
+    private function getProviderUserAttributes(
+        UserAttributeCollectionProviderInterface $userAttributeProvider, ?string $userIdentifier): array
+    {
+        // We cache the attributes for each provider, but only for the last userIdentifier
+        $providerKey = spl_object_id($userAttributeProvider);
+        if (false === array_key_exists($providerKey, $this->valuesPerCollectionProviderCache)
+            || $this->valuesPerCollectionProviderCache[$providerKey][0] !== $userIdentifier) {
+            $this->valuesPerCollectionProviderCache[$providerKey] = [$userIdentifier, $userAttributeProvider->getUserAttributes($userIdentifier)];
+        }
+        $cacheEntry = $this->valuesPerCollectionProviderCache[$providerKey];
+        assert($cacheEntry[0] === $userIdentifier);
+
+        return $cacheEntry[1];
+    }
+
+    /**
+     * @throws UserAttributeException
+     */
+    private function getProviderUserAttribute(
+        UserAttributeCollectionProviderInterface $allAvailableUserAttributesProvider,
+        ?string $userIdentifier, string $attributeName): mixed
+    {
+        $userAttributes = $this->getProviderUserAttributes($allAvailableUserAttributesProvider, $userIdentifier);
+        if (false === array_key_exists($attributeName, $userAttributes)) {
+            throw new UserAttributeException(sprintf('attribute \'%s\' was not provided', $attributeName),
+                UserAttributeException::USER_ATTRIBUTE_UNDEFINED);
+        }
+
+        return $userAttributes[$attributeName];
+    }
+
+    /**
+     * @return string[]
+     */
+    private function getAdditionallyAvailableAttributesCached(): array
+    {
+        if ($this->additionallyAvailableAttributesCache === null) {
+            $event = new GetAvailableUserAttributesEvent();
+            $this->eventDispatcher->dispatch($event);
+            $this->additionallyAvailableAttributesCache = $event->getAttributes();
+        }
+
+        return $this->additionallyAvailableAttributesCache;
     }
 }